move from utils to posteriors and return full posterior
