+++ sbi/inference/posteriors/mcmc_posterior.py
@@ -1,5 +1,6 @@
 # This file is part of sbi, a toolkit for simulation-based inference. sbi is licensed
 # under the Apache License Version 2.0, see <https://www.apache.org/licenses/>
+import inspect
 import warnings
 from copy import deepcopy
 from functools import partial
@@ -32,6 +33,7 @@ from sbi.samplers.mcmc import (
     sir_init,
 )
 from sbi.sbi_types import Shape, TorchTransform
+from sbi.utils import mcmc_transform
 from sbi.utils.potentialutils import pyro_potential_wrapper, transformed_potential
 from sbi.utils.torchutils import ensure_theta_batched, tensor2numpy
 
@@ -1110,3 +1112,80 @@ def _maybe_use_dict_entry(default: Any, key: str, dict_to_check: Dict) -> Any:
     """
     attribute = dict_to_check.get(key, default)
     return attribute
+
+
+def _num_required_args(func):
+    """
+    Utility for counting the number of positional args in a function.
+
+    This function counts each parameter in the signature that are positional -- ie.
+        (1) cannot only be passed in as keyword arguments
+        (2) do not have a default value
+
+    Args:
+        func: A callable function.
+
+    Returns:
+        Number of required positional arguments.
+    """
+    sig = inspect.signature(func)
+    return sum(
+        1
+        for param in sig.parameters.values()
+        if param.kind
+        in (inspect.Parameter.POSITIONAL_ONLY, inspect.Parameter.POSITIONAL_OR_KEYWORD)
+        and param.default is inspect._empty
+    )
+
+
+def build_from_potential(
+    potential_fn: Callable, prior: Any, x: Optional[Tensor] = None, **kwargs
+) -> MCMCPosterior:
+    """
+    Returns a sampler from a MCMCPosterior object, given user-defined potential
+    function and prior.
+
+    The user-defined potential can be conditional (accepts theta and x as positional
+    arguments) or unconditional (accepting only theta).
+
+    Args:
+        potential_fn: User defined potential function. Must be of type Callable.
+        prior: Prior distribution for parameter transformation and initialization.
+        x: Conditional x value. Provided if using a conditional potential function.
+
+    Returns:
+        Callable sampling function from MCMCPosterior object.
+    """
+    # build transformation to unrestricted space for sampling
+    transform = mcmc_transform(prior)
+
+    # potential_fn must take 1 or 2 required arguments: (theta) or (theta, x)
+    num_args = _num_required_args(potential_fn)
+    assert num_args > 0 and num_args < 3, (
+        "potential_fn must take 1-2 required arguments"
+    )
+    is_conditional = num_args == 2
+
+    if is_conditional:
+        # you could remove this and require use to set x before calling sample
+        assert x is not None, "x must be provided if potential_fn is conditional"
+        posterior = MCMCPosterior(potential_fn, prior, theta_transform=transform)
+        posterior.set_default_x(x)
+
+    else:
+        warn(
+            "x has not been provided. Using unconditional potential function.",
+            UserWarning,
+            stacklevel=2,
+        )
+
+        # define an unconditional potential function (ignores x)
+        def unconditional_potential_fn(theta, x):
+            return potential_fn(theta)
+
+        posterior = MCMCPosterior(
+            unconditional_potential_fn, prior, theta_transform=transform, **kwargs
+        )
+        posterior.set_default_x(torch.zeros(1))  # set default_x to dummy value
+
+    return posterior
